+++ src/client/client.h
@@ -203,10 +203,7 @@ extern clientActive_t cl;
  */
 typedef struct
 {
-	uint8_t frametime[MAX_TIMEDEMO_FRAMES];    ///< integer since frametimes aren't fractional
-	                                           ///< this should be a really low number always so unit8_t is more than enough,
-	                                           ///< if it isn't, your benchmark really isn't valid anyway
-
+	uint16_t frametime[MAX_TIMEDEMO_FRAMES];   ///< integer since frametimes aren't fractional
 	int timeFrames;                            ///< counter of rendered frames
 	int timeStart;                             ///< cls.realtime before first frame
 	int timeBaseTime;                          ///< each frame will be at this time + frameNum * 50
